made Sequence hash convert to uint64 before shift
I'm not clear on why the hash is shifted at all, but apparently it was important enough to be done twice to /make sure/. That said, the original implementation casted to uintptr_t, meaning hash results were different on different arches; this seems like a bad idea.

+++ sequence-immutable.go
@@ -767,7 +767,7 @@ func SequenceHash(vm *VM, target, locals Interface, msg *Message) Interface {
 	s := target.(*Sequence)
 	h := fnv.New32()
 	h.Write(s.Bytes())
-	return vm.NewNumber(float64(h.Sum32() << 2)) // ????????
+	return vm.NewNumber(float64(uint64(h.Sum32()) << 2)) // ????????
 }
 
 // SequenceInSlice is a Sequence method.
